requirements.txt
# requirements.txt — MedAssist AI v6.0 (Production-Ready)
# Install: pip install -r requirements.txt

# ─── Core AI Framework ────────────────────────────────────────────────────────
langgraph==0.2.64
langchain==0.3.14
langchain-core==0.3.30
langchain-community==0.3.14
langchain-groq==0.2.3
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
langgraph-checkpoint==2.0.10
langsmith==0.2.11

# ─── LLM Providers ────────────────────────────────────────────────────────────
groq==0.15.0
openai==1.59.8
tiktoken==0.8.0

# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn==0.34.0
starlette==0.41.3
python-multipart==0.0.12    # FIX 3: file uploads

# ─── Frontend ─────────────────────────────────────────────────────────────────
streamlit==1.41.1

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
PyMySQL==1.1.1
cryptography==43.0.3

# ─── FIX 1: JWT Authentication ────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4

# ─── Data Validation ──────────────────────────────────────────────────────────
pydantic==2.10.5
pydantic-settings==2.7.1
email-validator==2.2.0

# ─── HTTP & Networking ────────────────────────────────────────────────────────
requests==2.32.3
httpx==0.28.1

# ─── Search ───────────────────────────────────────────────────────────────────
tavily-python==0.5.0

# ─── Environment ──────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ─── Utilities ────────────────────────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.3
PyYAML==6.0.2
